9211 - quick refactor to have search functionality in its own function [skip ci]
